requirements.txt
Django==3.1
environs==8.0.0
Pillow==7.2.0
django-admin-sortable2==0.7.7